Initialize far
far is a Find And Replace command line program for replacing strings of
text within files. It aims to be simple and efficient.

The program currently supports finding and replacing in one file. This
commit does not contain tests, documentation or really anything beyond
a spike of getting _something_ somewhat working.

The goal is to have something that works like:

    far ./app MyClass NewClass

Future feature ideas:

- step through confirm or deny (like git add -p)
- output stats
- use .gitignore and/or .farignore for ignoring files and dirs

Go is chosen for the language to build this in as it seems solid for
building command line tools, and it is something that I am interested in
learning and using.