Restructure code and internal packages
Extract the code that was in main.go into two different packages: cmd &
far.

cmd is responsible for validating the user input and starting the
finding and replacing.

far is responsible for taking what cmd sends it and finding and
replacing the text.

This structure should be easier to manage (and hopefully easier to
test). It is based, somewhat off of the structure of
https://github.com/exercism/cli

My apologies if this is not very Go-like. :)